Update distance used to calculated point reached
Increase the distanced used to say that an agent has reached a point. If
the value is too low the agent will circle around the point without ever
'reaching' it and will consider itself stuck.

The distance should be hardcoded in the future as currently it is a
function of the agents size, meanig different size agents may experience
problems while others don't